% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{could_this_be_that}
\alias{could_this_be_that}
\alias{could_chr_be_num}
\alias{could_chr_be_int}
\alias{could_num_be_int}
\alias{could_chr_be_dtm}
\alias{could_dtm_be_dte}
\title{Tests is a vector could be of another data type}
\usage{
could_chr_be_num(.x)

could_chr_be_int(.x)

could_num_be_int(.x)

could_chr_be_dtm(.x)

could_dtm_be_dte(.x)
}
\arguments{
\item{.x}{vector of the data type that should be tested.}
}
\value{
TRUE or FALSE
}
\description{
Tests if vector could be a another data type without errors.
}
\details{
The name logic of \code{could_chr_be_num} should be interpreted as:
Could this character vector be a numeric vector?
The same logic goes for all functions named could_this_be_that.
}
\examples{
x <- c("1", "3", "7")
could_chr_be_num(x)
could_chr_be_int(x)

x <- c("abc", "3", "Hello world")
could_chr_be_num(x)

x <- c(NA, "3.45", "5,98")
could_chr_be_num(x)
could_chr_be_int(x)

x <- as.numeric(c(3.45, 1.5))
could_num_be_int(x)

x <- as.numeric(c(7, 2))
could_num_be_int(x)

}
