% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{create_dummy}
\alias{create_dummy}
\alias{dummy}
\alias{dummy_}
\title{Create a simple dummy}
\usage{
dummy(condition, missing = NA)
dummy_(condition, missing = 0L)

dummy_(condition, missing = 0L)
}
\arguments{
\item{condition}{a predicament}

\item{missing}{a replacement if condition is NA}
}
\value{
a vector of the integers 1, 0 and NA (if not dummy_ is used).
}
\description{
Creates a vector of the integers 1 and 0. If condition is true it returns 1. If false 0.
If condition returns NA it returns NA, if not explicitly not stated than NA
should be replaced.
}
\examples{
v <- c(10, 5, 3, NA, 9)
dummy(v > 5)
dummy_(v > 5)

}
\seealso{
\code{vignette("hablar")}
}
