% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{cumulative_}
\alias{cumulative_}
\alias{cumsum_}
\alias{cummean_}
\alias{cum_unique}
\alias{cum_unique_}
\title{cumulative_}
\usage{
cumsum_(.v, ignore_na = TRUE)

cummean_(.v, ignore_na = TRUE)

cum_unique(.v, ignore_na = FALSE)

cum_unique_(.v, ignore_na = TRUE)
}
\arguments{
\item{.v}{a vector}

\item{ignore_na}{should missing values be ignores?}
}
\value{
a vector
}
\description{
cumulative functions. 'cumsum_' is the cumulative sum ignoring missing values.
'cum_unique' counts the cumulative unique value including NA as ONE value. 
'cum_unique_' ignores missing values
}
