% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{given}
\alias{given}
\alias{given_}
\title{given}
\usage{
given(.x, .y, ignore_na = FALSE)

given_(.x, .y, ignore_na = TRUE)
}
\arguments{
\item{.x}{the vector to filter}

\item{.y}{a logical vector to filter with}

\item{ignore_na}{should NA be removed?}
}
\value{
a vector
}
\description{
Simple function that filters a vector while helping with missing values.
Replacing expression like 'x[x > 3 & !is.null(x)]'
}
