% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{math ignore NA in math funs}
\alias{math ignore NA in math funs}
\alias{\%minus_\%}
\alias{\%plus_\%}
\title{Ignore NA in math}
\usage{
.x \%minus_\% .y

.x \%plus_\% .y
}
\arguments{
\item{.x}{numeric or integer element}

\item{.y}{numeric or integer element}
}
\value{
a single value
}
\description{
Simplifying math functions are simple wrappers of math function (- +).
If any of the left-hand side or right-hand side is NA, Inf or NaN it 
returns any rational value, if there is any. 

However, if the both values are irriational it returns NA. 
The result is then passed to the
corresponding math function.
}
\examples{
# The simplest case
3 \%minus_\% 2

# But with NA it returns 3 as if the NA were zero
3 \%minus_\% NA

# It doesnt matter if the irrational number is on left- or right-hand.
NA \%plus_\% 5

}
\seealso{
\code{vignette("s")}, \code{vignette("hablar")}
}
