% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{n_unique count unique elements}
\alias{n_unique count unique elements}
\alias{n_unique}
\alias{n_unique_}
\title{n_unique}
\usage{
n_unique(.x, ignore_na = FALSE)

n_unique_(.x, ignore_na = TRUE)

n_unique_(.x, ignore_na = TRUE)
}
\arguments{
\item{.x}{a vector}

\item{ignore_na}{a logical indicating whether missing values should be removed}
}
\value{
a single numeric vector of the same length as the data frame it
is applied to.
}
\description{
Simple wrapper for length(unique(.x)). If you use n_unique_(.x) then NA
is ignored when counting.
}
\examples{

# Simple
n_unique(c(1, 2, 2, 3))

# Same result as above eventhough vector includes NA
n_unique_(c(1, 2, 2, 3, NA))

}
\seealso{
\code{vignette("s")}, \code{vignette("hablar")}
}
