% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{repeat_df}
\alias{repeat_df}
\title{repeat_df}
\usage{
repeat_df(.df, n, id = NULL)
}
\arguments{
\item{.df}{a data frame}

\item{n}{times the data frame should be repeated}

\item{id}{a character element that creates a column with a number for each repetition}
}
\value{
a vector of the integers 1, 0 and NA (if not dummy_ is used).
}
\description{
Repeats a data frame n times. Useful for testing on large data frames.
}
\examples{
repeat_df(mtcars, 2)

}
\seealso{
\code{vignette("hablar")}
}
