% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{replacers}
\alias{replacers}
\alias{if_na}
\alias{if_inf}
\alias{if_nan}
\alias{if_zero}
\alias{na_if}
\alias{inf_if}
\alias{nan_if}
\alias{zero_if}
\alias{if_not_na}
\title{replacemnt and specials}
\usage{
if_na(.x, replacement, missing = NULL)

if_nan(.x, replacement, missing = NULL)

if_inf(.x, replacement, missing = NULL)

if_zero(.x, replacement, missing = NULL)

na_if(.x, condition, replace_na = FALSE)

nan_if(.x, condition, replace_na = FALSE)

inf_if(.x, condition, replace_na = FALSE)

zero_if(.x, condition, replace_na = FALSE)

if_not_na(.x, replacement, missing = NULL)

if_inf(.x, replacement, missing = NULL)

if_nan(.x, replacement, missing = NULL)

if_zero(.x, replacement, missing = NULL)

na_if(.x, condition, replace_na = FALSE)

inf_if(.x, condition, replace_na = FALSE)

nan_if(.x, condition, replace_na = FALSE)

zero_if(.x, condition, replace_na = FALSE)
}
\arguments{
\item{.x}{a vector}

\item{replacement}{a replacement if condition is TRUE}

\item{missing}{a value that replace missing values in condition.}

\item{condition}{a predicament}

\item{replace_na}{if TRUE, missing values in condition will be replaced as well}
}
\value{
a vector
}
\description{
If-this-type-then replace with x. And the other way around; replace with x
if this.
}
\examples{

v <- c(1, NA, 2)
if_na(v, 100)

v <- c(999, NA, 2)
zero_if(v, v == 999)

}
\seealso{
\code{vignette("s")}, \code{vignette("hablar")}
}
