% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{set_wd_to_script_path}
\alias{set_wd_to_script_path}
\title{Set wd to script path}
\usage{
set_wd_to_script_path()
}
\value{
NULL. In the background the working directory has changed if not any errors occurred.
}
\description{
Sets working directory to the path where the R-script is located. Only works
inside [Rstudio] and in a script (i.e. not in the console). Additionally, the R-script needs to
be saved in a path to work.
}
