% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hablar.R
\name{this_date}
\alias{this_date}
\alias{this_day}
\alias{this_month}
\alias{this_year}
\title{this_date}
\usage{
this_day()

this_month()

this_year()
}
\value{
a date or number
}
\description{
Returns the current day, month or year. Day and month returns dates and year a 4 digit number.
}
\examples{
this_day() 
this_month() 
this_year() 

}
